Credential branding is an overridable configuration
