+++ core/App/contexts/configuration.tsx
@@ -1,12 +1,14 @@
 import { createContext, useContext } from 'react'
 
 import OnboardingPages from '../screens/OnboardingPages'
+import { Overlay } from '../types/overlay'
 
 export interface ConfigurationContext {
   pages: typeof OnboardingPages
   splash: React.FC
   terms: React.FC
   homeContentView: React.FC
+  OCABundle: Record<string, Overlay>
 }
 
 export const ConfigurationContext = createContext<ConfigurationContext>(null as unknown as ConfigurationContext)
